customer-marketing-ai-main/requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
torch
prefect
pyyaml
